quanstats_tearsheets/requirements.txt
appdirs==1.4.4
async-timeout==4.0.2
beautifulsoup4==4.11.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
contourpy==1.0.7
cryptography==39.0.1
cycler==0.11.0
fonttools==4.38.0
frozendict==2.3.5
html5lib==1.1
idna==3.4
kiwisolver==1.4.4
loguru==0.6.0
lxml==4.9.2
matplotlib==3.7.0
multitasking==0.0.11
mysql-connector-python==8.0.32
numpy==1.24.2
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
protobuf==3.20.3
pycparser==2.21
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
QuantStats==0.0.59
redis==4.5.1
requests==2.28.2
scipy==1.10.1
seaborn==0.12.2
six==1.16.0
soupsieve==2.4
tabulate==0.9.0
urllib3==1.26.14
webencodings==0.5.1
yfinance==0.2.12
